files added

+++ controllers/productsController.js
@@ -6,7 +6,7 @@ exports.listcategory = (req, res) => {
   const query = `SELECT * FROM category`;
   db.query(query, (err, rows) => {
     if (err) {
-      console.log(err);
+    
       res.status(500).send("Internal Server Error");
     } else {
       res.json(rows);
@@ -29,7 +29,7 @@ exports.listbrand = (req, res) => {
   `;
   db.query(query, (err, rows) => {
     if (err) {
-      console.log(err);
+    
       res.status(500).send("Internal Server Error");
     } else {
       res.json(rows);
@@ -56,7 +56,7 @@ FROM
 `;
   db.query(query, (err, rows) => {
     if (err) {
-      console.log(err);
+   
       res.status(500).send("Internal Server Error");
     } else {
       res.json(rows);
@@ -64,7 +64,7 @@ FROM
   });
 };
 
-//!                                SEARCH
+//!                             LIST SEARCH
 
 exports.search = (req, res) => {
   const { search } = req.body;
@@ -85,7 +85,7 @@ exports.search = (req, res) => {
 
   db.query(query, (err, rows) => {
     if (err) {
-      console.log(err);
+    
       res.status(500).send("Internal Server Error");
     } else {
       if (rows.length === 0) {
@@ -97,8 +97,7 @@ exports.search = (req, res) => {
   });
 };
 
-
-//!                             CATEGORY
+//!                           LIST CATEGORY
 
 exports.category = (req, res) => {
   const category = req.params.category;
@@ -118,7 +117,7 @@ FROM
 WHERE category_name = '${category}'`;
   db.query(query, (err, rows) => {
     if (err) {
-      console.log(err);
+  
       res.status(500).send("Internal Server Error");
     } else {
       //console.log(rows);
@@ -127,7 +126,7 @@ WHERE category_name = '${category}'`;
   });
 };
 
-//!                                  BRAND
+//!                                LIST BRAND
 
 exports.brand = (req, res) => {
   const brand = req.params.brand;
@@ -148,7 +147,7 @@ WHERE brand_name = '${brand}'`;
 
   db.query(query, (err, rows) => {
     if (err) {
-      console.log(err);
+    
       res.status(500).send("Internal Server Error");
     } else {
       //console.log(rows);
@@ -157,7 +156,7 @@ WHERE brand_name = '${brand}'`;
   });
 };
 
-//!                         BRAND AND CATEGORY
+//!                       LIST BRAND AND CATEGORY
 
 exports.categorybrand = (req, res) => {
   const brand = req.params.brand;
@@ -179,7 +178,7 @@ WHERE brand_name = '${brand}' AND category_name = '${category}'`;
 
   db.query(query, (err, rows) => {
     if (err) {
-      console.log(err);
+   
       res.status(500).send("Internal Server Error");
     } else {
       if (rows.length === 0) {
@@ -196,7 +195,7 @@ WHERE brand_name = '${brand}' AND category_name = '${category}'`;
 exports.addcategory = (req, res) => {
   try {
     if (!req.body) {
-      return res.status(400).json({ error: "Missing request body." });
+      return res.status(400).send("Missing request body.");
     }
 
     // check if category name is a string of characters
@@ -204,7 +203,7 @@ exports.addcategory = (req, res) => {
     if (!/^[A-Za-z]+$/.test(category_name)) {
       return res
         .status(400)
-        .json({ error: "Category name should contain only characters." });
+        .send("Category name should contain only characters.");
     }
 
     // check if category already exists
@@ -213,12 +212,12 @@ exports.addcategory = (req, res) => {
       [category_name],
       (err, rows) => {
         if (err) {
-          console.log(err);
+        
           res.send("Error checking for existing category.");
         } else {
           if (rows.length > 0) {
             // category already exists
-            res.status(400).json({ error: "Category already exists." });
+            res.status(400).send("Category already exists.");
           } else {
             // add category
             db.query(
@@ -228,7 +227,7 @@ exports.addcategory = (req, res) => {
                 if (!err) {
                   res.send("Category added successfully.");
                 } else {
-                  console.log(err);
+                
                   res.send("Error adding category.");
                 }
               }
@@ -238,8 +237,8 @@ exports.addcategory = (req, res) => {
       }
     );
   } catch (err) {
-    console.log(err);
-    res.status(500).json({ error: "Internal server error." });
+  
+    res.status(500).send("Internal server error.");
   }
 };
 
@@ -248,20 +247,16 @@ exports.addcategory = (req, res) => {
 exports.addbrand = (req, res) => {
   try {
     if (!req.body) {
-      return res.status(400).json({ error: "Missing request body." });
+      return res.status(400).send("Missing request body.");
     }
 
     // check if brand_name and category_id are strings of characters
     const { brand_name, category_id } = req.body;
     if (!/^[A-Za-z]+$/.test(brand_name)) {
-      return res
-        .status(400)
-        .json({ error: "Brand name should contain only characters." });
+      return res.status(400).send("Brand name should contain only characters.");
     }
     if (isNaN(category_id) || category_id <= 0) {
-      return res
-        .status(400)
-        .json({ error: "Category ID should contain only integers." });
+      return res.status(400).send("Category ID should contain only integers.");
     }
     // check if brand already exists
     db.query(
@@ -269,12 +264,12 @@ exports.addbrand = (req, res) => {
       [brand_name, category_id],
       (err, rows) => {
         if (err) {
-          console.log(err);
+        
           res.send("Error checking for existing brand.");
         } else {
           if (rows.length > 0) {
             // brand already exists
-            res.status(400).json({ error: "Brand already exists." });
+            res.status(400).send("Brand already exists.");
           } else {
             // add brand
             db.query(
@@ -284,7 +279,7 @@ exports.addbrand = (req, res) => {
                 if (!err) {
                   res.send("Brand added successfully.");
                 } else {
-                  res.status(400).json({ error: "incorrect category_id" });
+                  res.status(400).send("incorrect category_id");
                 }
               }
             );
@@ -293,8 +288,8 @@ exports.addbrand = (req, res) => {
       }
     );
   } catch (err) {
-    console.log(err);
-    res.status(500).json({ error: "Internal server error." });
+  
+    res.status(500).send("Internal server error.");
   }
 };
 
@@ -303,7 +298,7 @@ exports.addbrand = (req, res) => {
 exports.addproduct = (req, res) => {
   try {
     if (!req.body) {
-      return res.status(400).json({ error: "Missing request body." });
+      return res.status(400).send("Missing request body.");
     }
 
     const { product_name, product_quantity, product_brand, product_category } =
@@ -313,7 +308,7 @@ exports.addproduct = (req, res) => {
       typeof product_name !== "string" ||
       !/^[a-zA-Z ]+$/.test(product_name)
     ) {
-      throw new Error("Product name must contain only alphabets and spaces.");
+      res.send("Product name must contain only alphabets and spaces.");
     }
 
     if (
@@ -324,14 +319,14 @@ exports.addproduct = (req, res) => {
       typeof product_category !== "number" ||
       !Number.isInteger(product_category)
     ) {
-      throw new Error("Product quantity must be an integer.");
+      res.send("Product quantity must be an integer.");
     }
     db.query(
       "SELECT * FROM products WHERE product_name = ? AND product_brand = ? AND product_category = ?",
       [product_name, product_brand, product_category],
       (err, rows) => {
         if (err) {
-          console.log(err);
+        
           res.send("Error checking for existing product.");
         } else {
           if (rows.length > 0) {
@@ -341,11 +336,11 @@ exports.addproduct = (req, res) => {
               [newQuantity, rows[0].product_id],
               (err, rows) => {
                 if (err) {
-                  console.log(err);
+            
                   res.send("Error updating product quantity.");
                 } else {
                   res.send(
-                    "Product is already exit so, Product quantity updated successfully."
+                    "Product is already exit So, Product quantity updated successfully."
                   );
                 }
               }
@@ -358,7 +353,7 @@ exports.addproduct = (req, res) => {
                 if (!err) {
                   res.send("Product added successfully.");
                 } else {
-                  console.log(err);
+            
                   res.send("Error adding product.");
                 }
               }
@@ -368,7 +363,7 @@ exports.addproduct = (req, res) => {
       }
     );
   } catch (err) {
-    console.log(err);
+  
     res.status(400).json({ error: err.message });
   }
 };
@@ -380,7 +375,7 @@ exports.buy = (req, res) => {
   const quantity = req.body.quantity;
   try {
     if (isNaN(quantity) || quantity <= 0)
-      throw new Error("Please enter a valid number");
+      new Error("Please enter a valid number");
   } catch (err) {
     res.status(400).send(err.message);
     return;
@@ -439,7 +434,7 @@ exports.editcategory = (req, res) => {
   if (!/^[A-Za-z]+$/.test(category_name)) {
     return res
       .status(400)
-      .json({ error: "Category_name  should contain only characters." });
+      .send("Category_name  should contain only characters.");
   }
   db.query(
     `UPDATE category SET 
@@ -462,16 +457,13 @@ exports.editcategory = (req, res) => {
   );
 };
 
-
 //!                            EDIT BRAND
 
 exports.editbrand = (req, res) => {
   const brand_id = req.params.brand_id;
   const { brand_name } = req.body;
   if (!/^[A-Za-z]+$/.test(brand_name)) {
-    return res
-      .status(400)
-      .json({ error: "brand_name  should contain only characters." });
+    return res.status(400).send("brand_name  should contain only characters.");
   }
   db.query(
     `UPDATE brand SET 
@@ -498,7 +490,8 @@ exports.editbrand = (req, res) => {
 
 exports.editproduct = (req, res) => {
   const product_id = req.params.product_id;
-  const { product_name, product_quantity, product_category, product_brand } = req.body;
+  const { product_name, product_quantity, product_category, product_brand } =
+    req.body;
 
   // Check if product_id is a valid integer
   if (isNaN(parseInt(product_id))) {
@@ -507,8 +500,14 @@ exports.editproduct = (req, res) => {
   }
 
   // Check if product_quantity, product_category, and product_brand are valid integers and greater than 0
-  if (isNaN(parseInt(product_quantity)) || isNaN(parseInt(product_category)) || isNaN(parseInt(product_brand)) ||
-      parseInt(product_quantity) <= 0 || parseInt(product_category) <= 0 || parseInt(product_brand) <= 0) {
+  if (
+    isNaN(parseInt(product_quantity)) ||
+    isNaN(parseInt(product_category)) ||
+    isNaN(parseInt(product_brand)) ||
+    parseInt(product_quantity) <= 0 ||
+    parseInt(product_category) <= 0 ||
+    parseInt(product_brand) <= 0
+  ) {
     res.status(400).send("Invalid product quantity, category, or brand");
     return;
   }
@@ -547,13 +546,16 @@ exports.editproduct = (req, res) => {
 
 exports.deletecategoryid = (req, res) => {
   const category_id = req.params.id;
-
+  if (isNaN(parseInt(category_id))) {
+    res.status(400).send("Invalid category_id");
+    return;
+  }
   db.query(
     "DELETE FROM category WHERE category_id = ?",
     [category_id],
     (err, result) => {
       if (err) {
-        console.log(err);
+      
         res.send("Error deleting category.");
       } else if (result.affectedRows === 0) {
         res.status(404).send("category not found.");
@@ -568,13 +570,16 @@ exports.deletecategoryid = (req, res) => {
 
 exports.deletebrandid = (req, res) => {
   const brand_id = req.params.id;
-
+  if (isNaN(parseInt(brand_id))) {
+    res.status(400).send("Invalid brand_id");
+    return;
+  }
   db.query(
     "DELETE FROM brand WHERE brand_id = ?",
     [brand_id],
     (err, result) => {
       if (err) {
-        console.log(err);
+      
         res.send("Error deleting brand.");
       } else if (result.affectedRows === 0) {
         res.status(404).send("brand not found.");
@@ -589,13 +594,16 @@ exports.deletebrandid = (req, res) => {
 
 exports.deleteproductid = (req, res) => {
   const product_id = req.params.id;
-
+  if (isNaN(parseInt(product_id))) {
+    res.status(400).send("Invalid product_id");
+    return;
+  }
   db.query(
     "DELETE FROM products WHERE product_id = ?",
     [product_id],
     (err, result) => {
       if (err) {
-        console.log(err);
+      
         res.send("Error deleting product.");
       } else if (result.affectedRows === 0) {
         res.status(404).send("Product not found.");
